Add css for product details page & refactor ejs/js
The partials.js hadles all the partial files, and
all the ejs page templates are inside the template
folder
