--- src/pages/productDetails.ejs
@@ -1,13 +1,0 @@
-<section class="product-details-body">
-    <a href="/">Home</a>
-    <h1>Product</h1>
-    
-    <img class="" src="<%= product.image %>" alt="product image">
-    <div class="">
-        <p><%= product.title %></p>
-        <p><%= product.description %></p>
-        <p class="product-details-price"><%= "$" + formatPrice(product.price) %></p>
-    </div>
-
-    <button class="add-to-cart-button">Add to cart</button>
-</section>
