+++ src/partials.js
@@ -1,0 +1,7 @@
+import header from "./template/_partial/header.ejs"
+// import footer from "./pages/_partials/footer.ejs"
+
+export default {
+    header,
+    // footer
+}
