+++ src/styles/productDetails.css
@@ -1,0 +1,28 @@
+
+.product-detail-body {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+    max-width: 1024px;
+    gap: 10px;
+    padding: 30px;
+    background-color: var(--color-lt-gray);
+}
+
+.product-detail-container {
+    display: flex;
+    flex-direction: column;
+    width: 300px;
+    align-items: center;
+}
+
+.product-detail-info-wrapper {
+    display: flex;
+    flex-direction: column;
+    padding: 10px;
+    gap: 10px;
+}
+.product-detail-description {
+    font-size: 0.8rem;
+}
