+++ src/template/_partial/header.ejs
@@ -1,0 +1,3 @@
+<header>
+    <h1>Hi! i'm the header.</h1>
+</header>
