+++ src/template/home.ejs
@@ -1,3 +1,5 @@
+<%= partials.header() %>
+
 <div class="home-section container">
     <h1>home</h1>
     <a href="#/cart-page">Cart</a>
